pyproject.toml
[project]
name = "billexact"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = ["Flask>=3.0.0","PyYAML>=6.0"]

[build-system]
requires = ["setuptools>=64","wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["billexact*"]
